database/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
numpy==2.2.2
pandas==2.2.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2025.1
urllib3==2.2.1
